backend/requirements.txt
 Django>=5.0,<6.0
 djangorestframework>=3.14,<3.15
 djangorestframework-simplejwt>=5.3,<6.0
 django-cors-headers>=4.3,<5.0
 django-filter>=24.2,<25.0
 python-dotenv>=1.0,<1.1
 dj-database-url>=2.1,<3.0
 psycopg2-binary>=2.9,<3.0
 drf-spectacular>=0.27,<0.28
 supabase>=2.5,<3.0
 gunicorn>=21.2,<22.0
 whitenoise>=6.6,<7.0
